requirements.txt
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.5.2
boto3==1.26.14
botocore==1.29.14
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
cs50==9.2.2
dnspython==2.3.0
exceptiongroup==1.1.3
fastapi==0.104.1
Flask==2.2.2
Flask-Mobility==1.1.0
greenlet==2.0.1
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.1
psycopg2-binary==2.9.5
py3dns==3.2.1
pydantic==2.5.1
pydantic_core==2.14.3
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.44
sqlparse==0.4.3
starlette==0.27.0
style==1.1.0
termcolor==2.1.1
trycourier==4.4.0
twilio==7.15.3
typing_extensions==4.8.0
update==0.0.1
urllib3==1.26.12
validate-email-address==1
Werkzeug==2.2.2
zipp==3.10.0
